+++ src/main/scala/mesosphere/marathon/raml/ReadinessConversions.scala
@@ -17,8 +17,8 @@ trait ReadinessConversions {
       protocol = check.protocol.toRaml,
       path = check.path,
       portName = check.portName,
-      intervalSeconds = check.interval.toSeconds,
-      timeoutSeconds = check.timeout.toSeconds,
+      intervalSeconds = check.interval.toSeconds.toInt,
+      timeoutSeconds = check.timeout.toSeconds.toInt,
       httpStatusCodesForReady = check.httpStatusCodesForReady.to[Seq],
       preserveLastResponse = Some(check.preserveLastResponse)
     )
